[TP-11536] TODO 1: Major Refactor so that code is more maintainable
