+++ app/models/concerns/money_navigator/constants.rb
@@ -1,0 +1,6 @@
+module MoneyNavigator::Constants
+
+  MASK_ALL = '11'
+  MASK_SOME = '01'
+  MASK_NONE = '00'
+end
