+++ app/models/concerns/money_navigator/symbols.rb
@@ -1,0 +1,127 @@
+module MoneyNavigator::Symbols
+  #IMPORTANT: #We must enforce the hard limits of the system
+  #
+  #The masking only works if the question and answer flags are the same width
+  #This means there are the following hard limits in the system:
+  #- No more than FLAG_SIZE questions can be asked by the system
+  #- No Question can have more than FLAG_SIZE answers
+  #
+  #NB:
+  #- FLAG_SIZE represents the larger of [total number of questions] and [maximum number of answers per question]
+  #- The output of all computed/created flag values must be validated by validate_flag(flag)  method
+  #- Use FLAG_FORMAT to create bit strings from integers in line with FLAG_SIZE
+  #
+  #Changing the value of FLAG_SIZE is therefore enough should you require the system to handle
+  #a different number of questions/answers in line with the above notes
+  FLAG_SIZE = 16
+  #The flag is a zero left padded FLAG_SIZE long bit long binarry string
+  FLAG_FORMAT = "%0#{FLAG_SIZE}i"
+  FULL_FLAG_FORMAT = "%0#{FLAG_SIZE*FLAG_SIZE*2}i"
+  def self.validate_flag(flag)
+    #Can't really process anything if flags exist that we can not handle. Only alternative is to BOMB out
+    raise "Flag #{qn_hash[:flag]} for question #{qn_hash[:code]} (index: #{index}) has size: #{qn_hash[:flag].length}. Expected size: #{FLAG_SIZE}" if flag.length != FLAG_SIZE
+  end
+
+  EMPTY = 'EMPTY'
+  ALL = 'ALL'
+
+  #Set up the flags reference hash/lookup-table that will be used by the system
+  FLAGS = HashWithIndifferentAccess.new
+  FLAGS[EMPTY] = FLAG_FORMAT % '0'
+  FLAGS[ALL] = '1'*FLAG_SIZE
+  (0..FLAG_SIZE).to_a.each { |index| FLAGS[index.to_s] = FLAG_FORMAT % (2**(index)).to_s(2) }
+
+  #setup the question/answer hashes
+  #- QUESTIONS: An array of all the questions from the yml file enriched with flag information and with the code standadised to lowercase
+  #- QUESTIONS_HASH: hash of all the questions keyed on the question code
+  #- ANSWERS_HASH: hash containing all possible answer codes (keyed by code) and enriched with the respective flag information
+  QUESTIONS_HASH = HashWithIndifferentAccess.new
+  ANSWERS_HASH = HashWithIndifferentAccess.new
+
+
+  #Enrich a given question hashbag with flag information and downcased codes
+  #As a sideeffect the bag is updated into the QUESTIONS_HASH for faster code based lookup
+  def self.enrich_question(qn_hash)
+    index = /\d*$/.match(qn_hash[:code])[0]
+    qn_hash[:code].downcase!
+    qn_hash[:flag] = FLAGS[index]
+    validate_flag(qn_hash[:flag])
+    QUESTIONS_HASH[qn_hash[:code]] = qn_hash
+    qn_hash[:responses].each do |resp|
+      index = /\d*$/.match(resp[:code])[0]
+      resp[:code].downcase!
+      resp[:flag] = FLAGS[index]
+      validate_flag(resp[:flag])
+      ANSWERS_HASH[resp[:code]] = resp.slice(:code, :flag)
+    end
+    qn_hash[:responses].sort! {|r1, r2| r1[:sort_order].to_i <=> r2[:sort_order].to_i}
+
+    qn_hash
+  end
+
+  QUESTIONS = I18n.translate('money_navigator_tool.questions')
+    .map {| qn_hash | enrich_question(qn_hash)}
+    .sort { |q1, q2| q1[:sort_order].to_i <=> q2[:sort_order].to_i }
+
+  ANSWERS_HASH['EMPTY'] = {code: EMPTY, flag: FLAGS[EMPTY]}
+
+
+  #TODO: Might be a good idea to move these rules into the translation file though not sure if that'll
+  #be placing more in there than we want.
+  #- Not a good idea to keep the rules here and the text there... maintenance headache maintaining the header symbols in two places.
+  #- Not a good idea to move everything here... translation of headings into other languages will be required going forward.
+  #
+  #The data representation of the logic that triggers content being displayed
+  #Format:
+  # [
+  #   {
+  #     section_code: <Section Code>,
+  #     headings: [ <this is the list of headings and their associated content plus triggers as exampled below>
+  #       {
+  #         heading_code: <header code>,
+  #        content:
+  #        [<this is the list of content that can appear under a heading. Usually just one article >
+  #          {
+  #            triggers: [
+  #               < a list of the QA combinations whose state triggers display of the content >
+  #             ],
+  #             masks: [
+  #               <A list bitmasks that determins whether the triggers turn the header on or off
+  #               each list element must contain as many flags as there are triggers and if any one of these masks
+  #               matches the triggers then the content will be displayed>
+  #             ],
+  #             article: <The CMS URL of the content affected >
+  #         }
+  #       ]
+  #     }
+  #   }
+  # ]
+  #
+  # This is effectively a list of section rules.
+  # - Each section rule ismade up of a list of heading rules.
+  # - Each heading rule is made up of a list of content rules
+  # - Each content rule contains
+  #   - a list of triggers that produce a 1,0 state depending on whether they are activated or not by the answers (AND logic)
+  #   - a mask that is applied to the concatenated result states of the individual triggers (OR logic)
+  #   - the content to display if the mask agrees with the state of the triggers
+  # which in turn have heading rules
+  #
+  #NB. the length of the mask: value below should equal the length of the triggers though that requirement is not enforced for additional flexibility
+  #If it is shorter then the extra trigers are disregarded
+  #e.g with 2 elements in the trigger array generating '11' meaning they are both triggered.
+  #the following masks produce the specified content rule outcome
+  #- '1' - the content is displayed but only the first mask output is considered i.e (results the same for '11' and '10' output of the triggers)
+  #- '0' - the content is displayed only if the first trigger is not pulled irrespective of the state of the other i.e (results the same for '01' and '00' output of the triggers)
+  #- '10' - the content is displayed only if the first trigger is pulled and the second is not i.e (results are '10')
+  #- '01' - the content is displayed only if the second trigger is pulled and the first is not i.e (results are '01')
+  #- '11' - the content is displayed only if both triggers are pulled i.e (results are '11')
+  #
+  #TODO: Think of a less misleading form of reuse
+  MASK_ALL = '11'
+  MASK_SOME = '01'
+  MASK_NONE = '00'
+
+  #TODO move the sections into seperate files to make the rules easier to manage
+  CONTENT_RULES = MoneyNavigator::ContentRules.all
+end
+
