+++ features/factories/questions.rb
@@ -1,4 +1,4 @@
-include Symbols
+include MoneyNavigator::Symbols
 
 #TODO consider moving these methods into a module if needed for other fixtures.
 #Also to prevent polution of the global namespace (though not sure that's important given the way Factorybot is coded)
